Add "_serverByUdacity"-folder which is the repository provided by Udacity.com to provided a simple backend which simulates to save the created contactlist-entries.
In detail its a hardcoded nodejs-application which provides simple get/post methods to work with static data.

The original repository can be found here: https://github.com/udacity/reactnd-contacts-server2
I've added the code from the  repository to this repository so a user can work with the project without cloning multiple repositories and moreover to still have the code even when the other repository may get deleted or not.